[coq] Update to work with Coq and math-comp master.
Dear fourcolor authors, please find in this PR a proposal for the set
of changes needed in order to compile the development in Coq and
math-comp `master` branches.

Due to my unfamiliarity with the code I didn't investigate whether the
changes should be considered OK, or if they do reveal a bug upstream.